// src_backup_unpaid/types/finance.ts
export interface SyncSettings {
    enabled: boolean;
    type: 'local' | 'smb' | 'dropbox';
    localPath?: string;
    smbConfig?: {
        server: string;
        share: string;
        user: string;
        pass: string;
    };
    dropboxToken?: string;
    lastSync?: number;
}

export interface AppSettings {
    currency: string;
    language: string;
    theme: string;
    sync: SyncSettings;
}